pyproject.toml
[project]
name = "taraba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "mcp[cli]>=1.6.0",
    "python-fasthtml>=0.12.9",
    "uvicorn>=0.34.0",
]
